Changes to edmx and work on patient
Remade edmx of DB and continued working on users methods

+++ DoctorsOffice/User.cs
@@ -26,7 +26,7 @@ namespace DoctorsOffice
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Username { get; set; }
-        public int AMKA { get; set; }
+        public long AMKA { get; set; }
         public string Password { get; set; }
         public string userType { get; set; }
